NIFI-11705 Addressing review comments, adding SystemUtils and renaming packages
